eval_service/requirements.txt
opentelemetry-sdk==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-instrumentation-httpx>=0.61b0
opentelemetry-semantic-conventions>=0.61b0
openai>=1.0.0
boto3>=1.34.0
httpx>=0.27.0
